Buck: Don't depend transitively on dropwizard library
Dropwizard library is shipped in gerrit core and thus included in
gerrit.war. Because gerrit-server transitively depends on it, it
is also included in plugin API. It's always wrong to include library
in plugin.jar binary artifact that already included in gerrit.war.
There is possibility of versions collision, when different library
versions included in gerrit.war and plugin.jar, not to mention that
this unnesessary increases binary plugin artifact size.

Probably this dependency was added, to allow plugin to use dropwizard
library in tree build. The correct way to achieve this is to export
the library, as was done in: Idcda84cd.

This change partialy reverts I4a694cf5 and Ifa93527d.

Change-Id: I35f9762ca2ae4adedc9237eff70c69fa1e64c53b